* Quicktime videocapture engine:
  - Enumeration of and selection between different capture devices
    enabled. Still have to find a setup where i can actually get
    > 1 camera detected simultaneously. iSight seems to get disabled
    as soon as an external cam is attached.

  - Got rid of image row padding. Now the memory buffer is reformatted
    to remove the row padding before any further processing. Makes
    things simpler for external image processing code.

  - Now we support 1 channel luminance and 3 channel RGB by a combi
    of using a mono GWorld pixel format + some remapping of values
    in 1 channel mode, and some contraction/reformattting of the
    image buffer in RGB mode to strip the A channel from the RGBA
    input.

    --> Makes life easier for external image processing code.

* Cosmetic changes to libdc1394 engine as well.

* DownloadPsychtoolbox.m - Add location of svn client on OS/X Leopard to
  the default svn client search path.

* New MEX File: PsychCV

  PsychCV (Psych(C)omputer(V)ision is a container for miscellaneous
  C / OpenGL / GPU / OpenCV routines somehow loosely related to computer
  vision applications in vision science.

  It is linked against the free software OpenCV library V1.0.0 and needs
  that installed to work - for now. We may later on statically link against
  relevant libraries to make life easier. The current version is OS/X Intel
  only, compiles for other platforms will follow when things have stabilized
  a bit.

  Currently implemented is a heavily modified version of the computer vision
  based eyetracker "StarBurst" from the OpenEyes toolkit. This is half-finished
  code which needs lots more work...

  -> Pure backup commit of the PsychSourceGL/Source/Common/PsychCV/ code branch
     and the modified XCode project file.

* other small bits here and there...



git-svn-id: https://psychtoolbox-3.googlecode.com/svn/trunk@948 f5efbe48-b203-0410-9429-fc5e098e5a54

+++ Psychtoolbox/DownloadPsychtoolbox.m
@@ -208,6 +208,9 @@ function DownloadPsychtoolbox(flavor,targetdirectory,downloadmethod)
 %              select initial choie of download protocol to use, in order to allow to
 %              bypass misconfigured proxies and firewalls. Problem reported
 %              by Patrick Mineault.
+% 01/08/08 mk  On OS/X, add an additional search path for 'svn': /usr/bin/
+%              as per suggestion of Donald Kalar. Presumably, Apples Leopard ships
+%              with a svn client preinstalled in that location.
 
 % Flush all MEX files: This is needed at least on M$-Windows for SVN to
 % work if Screen et al. are still loaded.
@@ -298,14 +301,21 @@ else
 
     % Currently, we only know how to check this for Mac OSX.
     if isOSX
-        svnpath='/usr/local/bin/';
-        if exist('/usr/local/bin/svn','file')~=2
-            fprintf('The Subversion client "svn" is not in its expected\n');
-            fprintf('location "/usr/local/bin/svn" on your disk. Please \n');
-            fprintf('download and install the most recent Subversion client from:\n');
-            fprintf('web http://metissian.com/projects/macosx/subversion/ -browser\n');
-            fprintf('and then run %s again.\n',mfilename);
-            error('Subversion client is missing. Please install it.');
+        % Try OS/X 10.5 Leopard install location for svn first:
+        svnpath='/usr/bin/';
+        if exist('/usr/bin/svn','file')~=2
+            % This would have been the default install location of the svn
+            % client bundled with OS/X 10.5.x Leopard. Let's try the
+            % default install location from the web installer:
+            svnpath='/usr/local/bin/';
+            if exist('/usr/local/bin/svn','file')~=2
+                fprintf('The Subversion client "svn" is not in its expected\n');
+                fprintf('location "/usr/local/bin/svn" on your disk. Please \n');
+                fprintf('download and install the most recent Subversion client from:\n');
+                fprintf('web http://metissian.com/projects/macosx/subversion/ -browser\n');
+                fprintf('and then run %s again.\n',mfilename);
+                error('Subversion client is missing. Please install it.');
+            end
         end
     end
 end
